[resumes][refactor] Abstract comment edit form and votes to their components
